+++ packages/odata-service-writer/src/updates.ts
@@ -54,11 +54,17 @@ export function updatePackageJson(path: string, fs: Editor, addMockServer: boole
     }
 
     if (addMockServer) {
-        if (!packageJson.devDependencies['@sap/ux-ui5-fe-mockserver-middleware']) {
-            packageJson.devDependencies['@sap/ux-ui5-fe-mockserver-middleware'] = '1';
+        if (
+            !packageJson.devDependencies['@sap/ux-ui5-fe-mockserver-middleware'] &&
+            !packageJson.devDependencies['@sap-ux/ui5-middleware-fe-mockserver']
+        ) {
+            packageJson.devDependencies['@sap-ux/ui5-middleware-fe-mockserver'] = '2';
         }
-        if (!packageJson.ui5.dependencies.includes('@sap/ux-ui5-fe-mockserver-middleware')) {
-            packageJson.ui5.dependencies.push('@sap/ux-ui5-fe-mockserver-middleware');
+        if (
+            !packageJson.ui5.dependencies.includes('@sap/ux-ui5-fe-mockserver-middleware') &&
+            !packageJson.ui5.dependencies.includes('@sap-ux/ui5-middleware-fe-mockserver')
+        ) {
+            packageJson.ui5.dependencies.push('@sap-ux/ui5-middleware-fe-mockserver');
         }
     }
     fs.writeJSON(path, packageJson);